refactored create program/associate user
- moved createprogramtemp to a handler file program.go
- started changing function to rely on db functions instead of firebase
